% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils-deprecated.R
\name{boxplot_by_timeperiod}
\alias{boxplot_by_timeperiod}
\title{Deprecated: Returns a ggplot object giving a boxplot of the range of N 
within each time period.}
\usage{
boxplot_by_timeperiod(
  df,
  period_breaks,
  stat = "identity",
  scenario_labels,
  scenario_colors
)
}
\arguments{
\item{df}{a data frame with columns time, sim_num, scenario_name and N}

\item{period_breaks}{the dates to break up the display periods.}

\item{stat}{either "identity" or "peak" [TODO:extend options]}

\item{hosp_state_totals, }{}
}
\description{
Deprecated: Returns a ggplot object giving a boxplot of the range of N 
within each time period.
}
