% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{calcR0}
\alias{calcR0}
\title{Function calculates R from model outputs using R0 package}
\usage{
calcR0(
  USAfacts,
  geodat,
  incl_geoids,
  by_geoid = FALSE,
  min.date = NULL,
  max.date = NULL,
  pop_col = "pop"
)
}
\arguments{
\item{USAfacts}{df with observed cases col NincidConfirmed}

\item{geodat}{geodata file}

\item{incl_geoids}{geoids to include}

\item{by_geoid}{estimate R for each county}

\item{min.date}{start date for analysis}

\item{max.date}{end date for analysis}

\item{pop_col}{name of geodat column with population data}
}
\description{
Function calculates R from model outputs using R0 package
}
\author{
Kyra Grantz
}
