% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{forecast_plot}
\alias{forecast_plot}
\title{Forecast estimates compared to groundtruth and reichlab}
\usage{
forecast_plot(
  usa_facts = NULL,
  county_dat = res_state,
  forecast_start = projection_start,
  scenarios = forecast,
  reichlab = NULL,
  var = "NincidCase",
  pi_lo = 0.025,
  pi_hi = 0.975,
  truth_source = "JHU CSSE",
  color_vals = scn_colors,
  xmin_date = min_date,
  tendency = "mean"
)
}
\arguments{
\item{usa_facts}{df with groundtruth columns "incidI", "incidDeath" or "hosps"}

\item{county_dat}{df with incident cases, hospitalizations, and deaths}

\item{forecast_start}{last date of groundtruth}

\item{scenarios}{vector of scenarios to include}

\item{reichlab}{df with reichlab estimates from covidHubUtils::load_latest_forecast()}

\item{var}{name of variable in country_dat to plot}

\item{pi_lo}{lower limit to interval}

\item{pi_hi}{upper limit to interval}

\item{truth_source}{label for groundtruth}

\item{color_vals}{colors for scenarios and reichlab estimate}

\item{xmin_date}{start date for plot}

\item{tendency}{mean or median}

\item{geodat}{geodata file with geoid and name columns}
}
\description{
Forecast estimates compared to groundtruth and reichlab
}
