% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData-deprecated.R, R/ReportLoadData.R
\name{load_cum_hosp_geounit_date}
\alias{load_cum_hosp_geounit_date}
\title{Deprecated convenience function to load cumulative geounit hosp outcomes at a specific date for the given scenario}
\usage{
load_cum_hosp_geounit_date(
  outcome_dir,
  display_dates = config$report$formatting$display_dates,
  scenario_levels = NULL,
  scenario_labels = NULL,
  incl_geoids,
  pdeath_filter = "med",
  inference = TRUE
)

load_cum_hosp_geounit_date(
  outcome_dir,
  display_dates = config$report$formatting$display_dates,
  scenario_levels = NULL,
  scenario_labels = NULL,
  incl_geoids,
  pdeath_filter = "med",
  inference = TRUE
)
}
\arguments{
\item{scenario_labels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{incl_geoids}{character vector of geoids that are included in the report}

\item{pdeath_filter}{character string that filenames should match}

\item{scenariolabels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{name_filter}{character string that filenames should match}

\item{scn_dirs}{paste(config$name, config$interventions$scenarios, sep = "_") character vector of scenario directory names}

\item{display_date}{character date string for which cumulative infections should be extracted}

\item{geoid_len}{in defined, this we want to make geoids all the same length}

\item{padding_char}{character to add to the front of geoids if fixed length}
}
\value{
a data frame with columns
         - time
         - comp
         - geoid
         - N
         - sim_num
         - scenario_num
         - scenario_name

a data frame with columns
         - time
         - comp
         - geoid
         - N
         - sim_num
         - scenario
         - scenario_num
         - scenario_name
}
\description{
TODO : FIX ME
}
