% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_cum_inf_geounit_dates}
\alias{load_cum_inf_geounit_dates}
\title{Convenience function to load cumulative geounit infections at multiple specific dates for the given scenarios}
\usage{
load_cum_inf_geounit_dates(
  outcome_dir,
  varname = "incidI",
  display_dates = config$report$formatting$display_dates,
  scenario_levels = NULL,
  scenario_labels = NULL,
  incl_geoids,
  pdeath_filter = "med",
  inference = TRUE
)
}
\arguments{
\item{outcome_dir}{paste(config$name, config$interventions$scenarios, sep = "_") character vector of scenario directory names}

\item{varname}{with variable to download (e.g. 'incidI' or 'incidC')}

\item{display_dates}{config$report$formatting$display_dates character vector of dates to for which cumulative infections should be extracted}

\item{incl_geoids}{character vector of geoids that are included in the report}

\item{pdeath_filter}{string that indicates which pdeath level to import for the source files (from the hosp file name)}

\item{scenariolevels}{config$report$formatting$scenario_labels_short character vector of scenario levels}

\item{scenariolabels}{config$report$formatting$scenario_labels character vector of scenario labels}
}
\value{
a data frame with columns
         - time
         - geoid
         - sim_num
         - N
         - scenario_num
         - scenario_name
         - pdeath
}
\description{
Convenience function to load cumulative geounit infections at multiple specific dates for the given scenarios
}
