% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_hosp_geocombined_totals}
\alias{load_hosp_geocombined_totals}
\title{Convenience function to allow us to load hospital totals for the combined geounits quickly for
the given scenarios.}
\usage{
load_hosp_geocombined_totals(
  outcome_dir,
  scenario_levels,
  scenario_labels,
  pdeath_filter = c("high", "med", "low"),
  pre_process = function(x) {     x },
  incl_geoids,
  inference = TRUE,
  vacc_compartment = FALSE,
  ...
)
}
\arguments{
\item{outcome_dir}{the subdirectory with all model outputs}

\item{scenario_levels}{used to create scenario_name for labelling plots}

\item{scenario_labels}{used to create scenario_name for labelling plots}

\item{pdeath_filter}{string that indicates which pdeath to import from outcome_dir}

\item{pre_process}{function that does processing before collection}

\item{incl_geoids}{character vector of geoids that are included in the report}
}
\value{
a combined data frame of all hospital simulations with filters applied pre merge.
}
\description{
Convenience function to allow us to load hospital totals for the combined geounits quickly for
the given scenarios.
}
