% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData-deprecated.R
\name{load_hosp_geounit_peak_date}
\alias{load_hosp_geounit_peak_date}
\title{Deprecated convenience function to load peak geounit hospitalizations (or any other variable) by a specific date for the given scenarios}
\usage{
load_hosp_geounit_peak_date(
  scn_dirs,
  max_var,
  display_date = config$end_date,
  num_files = NA,
  name_filter,
  incl_geoids = NULL,
  scenariolabels = NULL,
  geoid_len = 0,
  padding_char = "0",
  file_extension = "auto"
)
}
\arguments{
\item{scn_dirs}{paste(config$name, config$interventions$scenarios, sep = "_") character vector of scenario directory names}

\item{max_var}{character string of variable that will be maximized per geoid}

\item{display_date}{date before which we should search for peaks}

\item{name_filter}{character string that filenames should match}

\item{incl_geoids}{optional character vector of geoids that are included in the report, if not included, all geoids will be used}

\item{scenariolabels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{geoid_len}{required length of geoid}

\item{padding_char}{padding}
}
\value{
a data frame with columns
        - sim_num
        - Pk_[variableName] variable that was maximized by geoid
        - NhospCurr number of people in hospital on a day
        - NICUCurr number of people in ICU on a day
        - NincidDeath  number of incidence deaths on a day
        - NincidInf  number of incident infections on a day
        - NincidICH number of incident ICUs on a day
}
\description{
Deprecated convenience function to load peak geounit hospitalizations (or any other variable) by a specific date for the given scenarios
}
