% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_hosp_geounit_relative_to_threshold}
\alias{load_hosp_geounit_relative_to_threshold}
\title{Convenience function to display log proportion}
\usage{
load_hosp_geounit_relative_to_threshold(
  outcome_dir,
  threshold,
  variable,
  scenario_levels,
  scenario_labels,
  end_date = config$end_date,
  incl_geoids = NULL,
  pdeath_filter,
  geodat = geodata,
  inference = TRUE,
  week = FALSE
)
}
\arguments{
\item{outcome_dir}{df with disaggregated hosp outcomes}

\item{threshold}{A named numeric vector where the names are geoids and values are thresholds}

\item{variable}{character string of variable to which to compare threshold}

\item{scenario_levels}{config$report$formatting$scenario_labels_short character vector of scenario levels}

\item{scenario_labels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{end_date}{simulation end date character string}

\item{incl_geoids}{optional character vector of geoids that are included in the report, if not included, all geoids will be used}

\item{pdeath_filter}{pdeath character string one of: high, med, low}

\item{inference}{passed on to load_hosp_sims}

\item{week}{will return an estimate for the weekly log_prop_needed peak of needed beds/ventilators}
}
\value{
a data frame with columns
        - scenario_name
        - geoid
        - time
        - NhospCurr number of people in hospital on a day
        - NICUCurr number of people in ICU on a day
        - NVentCurr  number of ventilators used on a day
        - threshold_value numeric values from threshold named vector
        - prop_needed ratio of needed beds/ventilators relative to threshold value
        - log_prop_needed log ratio of needed beds/ventilators with plotting edits on the borders
}
\description{
Convenience function to display log proportion
}
