% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_hosp_geounit_threshold}
\alias{load_hosp_geounit_threshold}
\title{Convenience function to load the slice for each geoid where the value of an outcome exceeds a given threshold}
\usage{
load_hosp_geounit_threshold(
  threshold,
  variable,
  end_date = config$end_date,
  pdeath_filter = "high",
  incl_geoids,
  scenario_labels,
  scenario_levels,
  outcome_dir,
  inference = TRUE
)
}
\arguments{
\item{threshold}{A named numeric vector.  This function will pull the first time slice that meets or exceeds all thresholds}

\item{variable}{character string of variable to which to compare threshold}

\item{end_date}{simulation end date character string}

\item{pdeath_filter}{pdeath selection one of: high, med, low}

\item{incl_geoids}{optional character vector of geoids that are included in the report, if not included, all geoids will be used}

\item{scenario_labels}{string with scenario levels passed on to load_hosp_totals}

\item{scenario_levels}{string with scenario names passed on to load_hosp_totals}
}
\value{
a data frame with columns
        - scenario_name
        - sim_num
        - NhospCurr number of people in hospital on a day
        - NICUCurr number of people in ICU on a day
        - NincidDeath  number of incidence deaths on a day
        - NincidInf  number of incident infections on a day
        - NincidICH number of incident ICUs on a day
}
\description{
Convenience function to load the slice for each geoid where the value of an outcome exceeds a given threshold
}
