% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_hosp_sims_filtered}
\alias{load_hosp_sims_filtered}
\title{Wrapper function for loading multiple hospitalization outcomes with open_dataset}
\usage{
load_hosp_sims_filtered(
  outcome_dir,
  model_output = "hosp",
  partitions = c("location", "scenario", "pdeath", "runID", "lik_type", "is_final",
    "filename"),
  pdeath_filter = c("high", "med", "low"),
  incl_geoids,
  pre_process = function(x) {     x },
  post_process = function(x) {     x },
  inference = TRUE,
  ...
)
}
\arguments{
\item{outcome_dir}{the subdirectory with all model outputs}

\item{model_output}{folder with hosp outcomes}

\item{partitions}{used by open_dataset}

\item{pdeath_filter}{string that indicates which pdeath to import from outcome_dir}

\item{incl_geoids}{character vector of geoids that are included in the report}

\item{pre_process}{function that does processing before collection}

\item{post_process}{function that does processing after collection}
}
\value{
a combined data frame of all hospital simulations with filters applied pre merge.
}
\description{
Wrapper function for loading multiple hospitalization outcomes with open_dataset
}
