% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_hpar_sims_filtered}
\alias{load_hpar_sims_filtered}
\title{Wrapper function for loading final hpar files with open_dataset}
\usage{
load_hpar_sims_filtered(
  outcome_dir,
  model_output = "hpar",
  partitions = c("location", "scenario", "pdeath", "runID", "lik_type", "is_final",
    "file_name"),
  pdeath_filter = c("high", "med", "low"),
  pre_process = function(x) {     x },
  incl_geoids,
  ...
)
}
\arguments{
\item{outcome_dir}{the subdirectory with all model outputs}

\item{model_output}{folder with hpar outcomes}

\item{partitions}{used by open_dataset}

\item{pdeath_filter}{string that indicates which pdeath to import from outcome_dir}

\item{pre_process}{function that does processing before collection}

\item{incl_geoids}{character vector of geoids that are included in the report}
}
\value{
a combined data frame of all hpar simulations with filters applied pre merge.
}
\description{
Wrapper function for loading final hpar files with open_dataset
}
