% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData-deprecated.R
\name{load_inf_geounit_peaks_date}
\alias{load_inf_geounit_peaks_date}
\title{Deprecated convenience function to load peak geounit infections before a given date for the given scenarios}
\usage{
load_inf_geounit_peaks_date(
  scn_dirs,
  display_date = config$end_date,
  num_files = NA,
  scenariolabels = NULL,
  incl_geoids = NULL,
  geoid_len = 0,
  padding_char = "0",
  file_extension = "auto"
)
}
\arguments{
\item{scn_dirs}{paste(config$name, config$interventions$scenarios, sep = "_") character vector of scenario directory names}

\item{display_date}{character string for date before which infection peak should be identified}

\item{scenariolabels}{config$report$formatting$scenario_labels character vector of scenario labels}

\item{incl_geoids}{optional character vector of geoids that are included in the report, if not included, all geoids will be used}

\item{geoid_len}{required length of geoid}

\item{padding_char}{padding}
}
\value{
a data frame with columns
         - time
         - comp
         - geoid
         - N
         - sim_num
         - scenario_num
         - scenario_name
}
\description{
Deprecated convenience function to load peak geounit infections before a given date for the given scenarios
}
