% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_jhu_csse_for_report}
\alias{load_jhu_csse_for_report}
\title{Load JHU CSSE data}
\usage{
load_jhu_csse_for_report(
  jhu_data_dir = "JHU_CSSE_Data",
  countries = c("US"),
  states
)
}
\arguments{
\item{jhu_data_dir}{data directory}

\item{countries}{character vector of countries}

\item{states}{character vector of states (state abbreviations is US-only)}
}
\value{
a data frame with columns
        - date
        - NcumulConfirmed
        - NcumulDeathsObs
        - NincidConfirmed
        - NincidDeathsObs
}
\description{
Load JHU CSSE data
}
