% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_npi_sims_filtered}
\alias{load_npi_sims_filtered}
\title{Convenience function to load intervention effectiveness estimates}
\usage{
load_npi_sims_filtered(
  outcome_dir,
  pdeath_filter = c("high", "med", "low"),
  incl_geoids,
  ...
)
}
\arguments{
\item{outcome_dir}{the subdirectory with all model outputs}

\item{pdeath_filter}{string that indicates which pdeath(s) to import from outcome_dir}

\item{incl_geoids}{character vector of geoids that are included in the report}
}
\value{
a combined data frame of daily Rt and total effectiveness estimates per geoid/sim
}
\description{
Convenience function to load intervention effectiveness estimates
}
