% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_r_daily_sims_filtered}
\alias{load_r_daily_sims_filtered}
\title{Convenience function for loading daily Rt and total effectiveness estimates by geoid}
\usage{
load_r_daily_sims_filtered(
  outcome_dir,
  pdeath_filter = c("high", "med", "low"),
  incl_geoids,
  mtr = TRUE,
  n_periods = 10,
  ...
)
}
\arguments{
\item{outcome_dir}{the subdirectory with all model outputs}

\item{pdeath_filter}{string that indicates which pdeath(s) to import from outcome_dir}

\item{incl_geoids}{character vector of geoids that are included in the report}

\item{mtr}{logical to determine if non-contiguous interventions were applied - default TRUE}

\item{n_periods}{maximum number of non-contiguous periods for any geoid/intervention}
}
\value{
a combined data frame of daily Rt and total effectiveness estimates per geoid
}
\description{
Convenience function for loading daily Rt and total effectiveness estimates by geoid
}
