% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_r_sims_filtered}
\alias{load_r_sims_filtered}
\title{Convenience function for loading intervention effect and R estimates}
\usage{
load_r_sims_filtered(
  outcome_dir,
  pdeath_filter = c("high", "med", "low"),
  pre_process = function(x) {     x },
  geodat = geodata,
  incl_geoids,
  npi_trimmer = "[[A-Z]].+\\\\_",
  ...
)
}
\arguments{
\item{outcome_dir}{the subdirectory with all model outputs}

\item{pdeath_filter}{string that indicates which pdeath(s) to import from outcome_dir}

\item{pre_process}{function that does processing before collection of snpi outputs}

\item{geodat}{df with geoid and name columns}

\item{incl_geoids}{character vector of geoids that are included in the report}

\item{npi_trimmer}{pattern used by str_remove to edit npi_name column; original 
names conserved in npi_group_name col}
}
\value{
a combined data frame of all R and effectiveness per geoid-intervention across 
simulations with filters applied pre merge
}
\description{
Convenience function for loading intervention effect and R estimates
}
