% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_seed_sims_filtered_interm}
\alias{load_seed_sims_filtered_interm}
\title{Wrapper function for loading intermediate seeding files from csv files}
\usage{
load_seed_sims_filtered_interm(
  outcome_dir,
  partitions = c("location", "scenario", "pdeath", "runID", "lik_type", "is_final",
    "file_name"),
  pdeath_filter = c("high", "med", "low"),
  incl_geoids,
  ...
)
}
\arguments{
\item{outcome_dir}{the subdirectory with all model outputs}

\item{partitions}{used by open_dataset}

\item{pdeath_filter}{string that indicates which pdeath to import from outcome_dir}

\item{incl_geoids}{character vector of geoids that are included in the report}
}
\value{
a combined data frame of all seeding values for all simulations with filters applied pre merge.
       - geoid
       - amount
       - scenario
       - pdeath
       - date
       - lik_type
       - is_final
       - slot_num
       - block_num
        -iter_num
}
\description{
Wrapper function for loading intermediate seeding files from csv files
}
