% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs.R
\name{load_spar_sims_filtered}
\alias{load_spar_sims_filtered}
\title{Wrapper function for loading spar files with open_dataset}
\usage{
load_spar_sims_filtered(
  outcome_dir,
  partitions = c("location", "scenario", "pdeath", "runID", "lik_type", "is_final",
    "filename"),
  pdeath_filter = c("high", "med", "low"),
  pre_process = list(partitions = function(x) {     x }, data = function(x) {     x }),
  ...
)
}
\arguments{
\item{outcome_dir}{the subdirectory with all model outputs}

\item{partitions}{used by open_dataset}

\item{pdeath_filter}{string that indicates which pdeath to import from outcome_dir}

\item{pre_process}{function that does processing before collection}
}
\value{
a combined data frame of all R simulations with filters applied pre merge.
       - parameter
       - value
       - location 
       - scenario
       - pdeath
       - date
       - lik_type
       - is_final
       - filename
}
\description{
Wrapper function for loading spar files with open_dataset
}
