% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportLoadData.R
\name{load_USAFacts_for_report}
\alias{load_USAFacts_for_report}
\title{Load USAFacts data either summarized or by geoid}
\usage{
load_USAFacts_for_report(
  data_dir = "data/case_data",
  incl_geoids = NULL,
  geodat = geodata,
  aggregate = FALSE
)
}
\arguments{
\item{data_dir}{data directory to download raw USAFacts data to}

\item{incl_geoids}{geoids to include if data will not be summarized}

\item{geodat}{df with geoid, name and population}

\item{aggregate}{whether to aggregate results by source}
}
\value{
a data frame with columns
        - date
        - NcumulConfirmed (Confirmed if not summarize)
        - NcumulDeathsObs (Deaths if not summarized)
        - NincidConfirmed (incidI if not summarized)
        - NincidDeathsObs (deaths)
}
\description{
Load USAFacts data either summarized or by geoid
}
