% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_CI}
\alias{make_CI}
\title{Function to print CIs}
\usage{
make_CI(lo, hi)
}
\arguments{
\item{lo}{numeric or vector of lower bound of CI}

\item{hi}{numeric or vector of upper bound of CI}
}
\value{
formatted texted for CI: '\(lo, hi\)'
}
\description{
Function to print CIs
}
