% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_scn_county_table_withVent}
\alias{make_scn_county_table_withVent}
\title{Function makes a summary table for each county}
\usage{
make_scn_county_table_withVent(
  current_scenario,
  county_dat,
  pi_lo = 0.025,
  pi_hi = 0.975,
  geodat = geodata,
  start_date,
  end_date,
  pdeath_filter = pdeath_default
)
}
\arguments{
\item{current_scenario}{scenario to summarize}

\item{county_dat}{contains the relevant hospital data}

\item{start_date}{summarization period start}

\item{end_date}{summarization period end}

\item{pdeath_filter}{if summarizing results for one pdeath only; leave NA to show all}

\item{pi_low}{low side of the prediction interval}

\item{pi_high}{high side of the prediction interval}

\item{pdeath_labels}{to label pdeath}

\item{pdeath_levels}{to order pdeaths}
}
\description{
Function makes a summary table for each county
}
