% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils-deprecated.R
\name{make_scn_state_table}
\alias{make_scn_state_table}
\title{Deprecated: Make statewide table of infections, hosp, ICU, deaths for given scenario. There is a function now make_scn_state_table_withVent that replaces this one.}
\usage{
make_scn_state_table(
  current_scenario,
  hosp_state_totals,
  table_dates,
  pdeath_labels,
  pdeath_filecode
)
}
\arguments{
\item{current_scenario}{text string of scenario label for which to build table}

\item{hosp_state_totals}{totals for hospitalization related data for state for all pdeath}

\item{table_dates}{formatted table_dates object}

\item{pdeath_labels}{pdeath formatted labels}

\item{pdeath_filecode}{pdeath file column codes}
}
\value{
state scenario table
}
\description{
Deprecated: Make statewide table of infections, hosp, ICU, deaths for given scenario. There is a function now make_scn_state_table_withVent that replaces this one.
}
