% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_scn_state_table_withVent}
\alias{make_scn_state_table_withVent}
\title{Make statewide table of infections, hosp, ICU, deaths, vents for given scenario}
\usage{
make_scn_state_table_withVent(
  current_scenario,
  hosp_state_totals,
  table_dates,
  pdeath_labels = c("1\%", "0.5\%", "0.25\%"),
  pdeath_levels = c("high", "med", "low")
)
}
\arguments{
\item{current_scenario}{text string of scenario label for which to build table}

\item{hosp_state_totals}{totals for hospitalization related data for state for all pdeath}

\item{table_dates}{formatted table_dates object}

\item{pdeath_labels}{pdeath formatted labels}

\item{pdeath_levels}{pdeath file column codes}
}
\value{
state scenario table
}
\description{
Make statewide table of infections, hosp, ICU, deaths, vents for given scenario
}
