% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_scn_time_summary_table_withVent}
\alias{make_scn_time_summary_table_withVent}
\title{Function makes a summary table for an entire state.}
\usage{
make_scn_time_summary_table_withVent(
  hosp_state_totals,
  period_breaks,
  pi_low = 0.025,
  pi_high = 0.975,
  round_digit = -2
)
}
\arguments{
\item{hosp_state_totals}{contains the relevant hospital data}

\item{period_breaks}{the dates to break up the display periods.}

\item{pi_low}{low side of the prediction interval}

\item{pi_high}{high side of the prediction interval}

\item{round_digit}{what level to round to}
}
\description{
Function makes a summary table for an entire state.
}
