% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_sparkline_tab_intervention_effect}
\alias{make_sparkline_tab_intervention_effect}
\title{Sparkline table for intervention period effectiveness estimates}
\usage{
make_sparkline_tab_intervention_effect(
  r_dat,
  npi_labels,
  npi_levels,
  current_scenario,
  pdeath_filter = pdeath_default,
  pi_lo = 0.025,
  pi_hi = 0.975,
  geodat = geodata,
  px_qual = 40,
  wh_ratio = 3.5,
  brewer_palette = "Spectral"
)
}
\arguments{
\item{r_dat}{df with reduction estimates per sim (from load_r_sims_filtered)}

\item{npi_labels}{labels for plotted NPIs}

\item{npi_levels}{levels of NPIs}

\item{current_scenario}{name of current scenario to use}

\item{pi_lo}{lower quantile for summarization}

\item{pi_hi}{higher quantile for summarization}

\item{geodat}{df with location names}

\item{px_qual}{sparkline pixel size passed on to ggplot_image}

\item{wh_ratio}{sparkline width:height ratio passed on to ggplot_image}

\item{brewer_palette}{pallete name passed on to brewer.pal}

\item{periodcolors}{}
}
\value{
a table with the effectiveness per intervention period and a bar graph
}
\description{
Sparkline table for intervention period effectiveness estimates
}
