% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{make_sparkline_tab_r}
\alias{make_sparkline_tab_r}
\title{Sparkline table with R estimates}
\usage{
make_sparkline_tab_r(
  r_dat,
  county_dat = NULL,
  susceptible = TRUE,
  current_scenario,
  npi_labels,
  npi_levels,
  pdeath_filter = pdeath_default,
  pi_lo = 0.025,
  pi_hi = 0.975,
  geodat = geodata,
  px_qual = 40,
  wh_ratio = 3.5,
  brewer_palette = "Spectral"
)
}
\arguments{
\item{r_dat}{df with R or reduction estimates per sim (only one scenario)}

\item{county_dat}{df with SEIR outputs}

\item{susceptible}{whether to show estimates adjusted for cumulative infections, 
in which case the estimate reflects the median Rt for past interventions and the Rt for
Sys.Date() for ongoing interventions}

\item{current_scenario}{name of scenario inputs to use}

\item{npi_labels}{labels for plotted NPIs}

\item{npi_levels}{levels of NPIs}

\item{pdeath_filter}{which pdeath value to select, does not support multiple pdeath}

\item{pi_lo}{lower quantile for summarization}

\item{pi_hi}{upper quantile for summarization}

\item{geodat}{df with location names}

\item{px_qual}{sparkline pixel size passed on to ggplot_image}

\item{wh_ratio}{sparkline width:height ratio passed on to ggplot_image}

\item{brewer_palette}{pallete name passed on to brewer.pal}
}
\value{
a table with the R per intervention period and a bar graph - note
}
\description{
Sparkline table with R estimates
}
