% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{mtr_estimates}
\alias{mtr_estimates}
\title{Function to add one row per intervention period for interventions using the MultiTimeReduce template}
\usage{
mtr_estimates(rt_dat, n_periods = 10)
}
\arguments{
\item{rt_dat}{df with estimates of snpi/spar outputs, with columns for geoid, scenario, pdeath, unique intervention names (must start with "npi"), start_date, end_date,}

\item{n_periods}{maximum number of non-contiguous dates}
}
\description{
Function to add one row per intervention period for interventions using the MultiTimeReduce template
}
