% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_accept_by_location_cumul}
\alias{plot_accept_by_location_cumul}
\title{Plot cumulative # of acceptances over MCMC iterations for each geoID and combined global}
\usage{
plot_accept_by_location_cumul(
  llik_interm,
  filter_by = "pdeath",
  filter_val = "med",
  burn_in = 0
)
}
\arguments{
\item{llik_interm}{df with geoid, name, ll, lik_type, scenario, pdeath, iter_num}

\item{filter_by}{variable name for filtering estimates either: scenario or pdeath}

\item{filter_val}{desired value of variable}

\item{burn_in}{number of iterations to discard at the beginning}
}
\value{
plot of chimeric and global acceptance rate by geoid
}
\description{
Plot cumulative # of acceptances over MCMC iterations for each geoID and combined global
}
