% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_county_outcomes}
\alias{plot_county_outcomes}
\title{Time series for cases, hospitalizations, and ICU by county and either pdeath or scenario}
\usage{
plot_county_outcomes(
  county_dat,
  filter_by = "pdeath",
  filter_val = pdeath_default,
  pi_lo = 0.025,
  pi_hi = 0.975,
  start_date,
  end_date,
  geodat = geodata,
  var_levels,
  var_labels
)
}
\arguments{
\item{county_dat}{df with incident cases, hospitalizations, and deaths}

\item{filter_by}{specify either one pdeath or scenario}

\item{filter_val}{value for filtering}

\item{pi_lo}{lower limit to interval}

\item{pi_hi}{upper limit to interval}

\item{start_date}{x-axis plot limits}

\item{end_date}{x-axis plot limits}

\item{geodat}{geodat file with name column}

\item{var_levels}{levels of comparison var}

\item{var_labels}{labels of comparison var}
}
\description{
Time series for cases, hospitalizations, and ICU by county and either pdeath or scenario
}
