% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_event_time_by_geoid}
\alias{plot_event_time_by_geoid}
\title{Plot figure showing when event time by geoid}
\usage{
plot_event_time_by_geoid(
  hosp_county_peaks,
  scenariolabel,
  shapefile,
  scenario_colors,
  time_caption,
  geoid_caption,
  value_name,
  value_label,
  start_date,
  end_date
)
}
\arguments{
\item{shapefile}{object with geoid and name}

\item{scenario_colors}{character vector of colors from config}

\item{time_caption}{label for time axis}

\item{geoid_caption}{label for geoid axis}

\item{value_name}{name of secondary axis value}

\item{value_label}{secondary axis character label}

\item{start_date}{start date as character string "2020-01-01"}

\item{end_date}{end date as character string}

\item{hosp_county_peak}{hosp geounit peak data}

\item{exclude_zeroes}{logical indicating whether to exclude geounits with no beds}
}
\value{
plot state time series median and IQR
}
\description{
Plot figure showing when event time by geoid
}
