% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_geounit_attack_rate_map}
\alias{plot_geounit_attack_rate_map}
\title{Plot map showing infections per 10K on a specific date for one scenario}
\usage{
plot_geounit_attack_rate_map(
  cum_inf_geounit_dates,
  geodat,
  shp,
  varlabel = "Infections",
  scenariolabel = config$report$formatting$scenario_labels[1],
  popnodes = "pop",
  display_date,
  viridis_palette = "plasma"
)
}
\arguments{
\item{cum_inf_geounit_dates}{dataframe with cumulative infections up through a specific date, produced by load_cum_inf_geounit_dates, perhaps}

\item{geodat}{as loaded by skeleton}

\item{shp}{shapefile with geounits}

\item{varlabel}{whether showing 'Infections' or 'Cases'}

\item{scenariolabel}{scenario label character string}

\item{popnodes}{name of pop variable in geodat}

\item{viridis_palette}{character string of viridis palette}

\item{display_datecharacter}{string of display date for map}
}
\value{
plot of cumulative infections per 10K by a specific date by geounit for a single scenario
}
\description{
Plot map showing infections per 10K on a specific date for one scenario
}
