% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils-deprecated.R,
%   R/ReportBuildUtils.R
\name{plot_geounit_map}
\alias{plot_geounit_map}
\title{Plot map showing chosen variable per 10K on a specific date for one scenario (blue/green color scale)}
\usage{
plot_geounit_map(
  cum_inf_geounit_dates,
  plot_var,
  geodat,
  shp,
  legend_name = "Value per 10K",
  scenariolabel = config$report$formatting$scenario_labels[1],
  popnodes = "pop",
  display_date,
  clims = NULL
)

plot_geounit_map(
  cum_inf_geounit_dates,
  plot_var,
  geodat,
  shp,
  legend_name = "Value per 10K",
  scenariolabel = config$report$formatting$scenario_labels[1],
  popnodes = "pop",
  display_date,
  clims = NULL
)
}
\arguments{
\item{cum_inf_geounit_dates}{dataframe produced by load_cum_inf_geounit_dates, with cumulative totals/averages on a given date}

\item{plot_var}{name of the variable of which to plot intensity}

\item{geodat}{as loaded by skeleton}

\item{shp}{shapefile with geounits}

\item{scenariolabel}{scenario label character string}

\item{popnodes}{name of pop variable in geodat}

\item{display_date}{string of display date for map}

\item{geodata}{as loaded by skeleton}

\item{viridis_palette}{character string of viridis palette}
}
\value{
map by geounit, filled by variable per 10K at a specific date for a single scenario

map by geounit, filled by variable per 10K at a specific date for a single scenario
}
\description{
Plot map showing chosen variable per 10K on a specific date for one scenario (blue/green color scale)

Plot map showing chosen variable per 10K on a specific date for one scenario (blue/green color scale)
}
