% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_hnpi_by_location}
\alias{plot_hnpi_by_location}
\title{Plot intermediate hnpi over time for each geoID}
\usage{
plot_hnpi_by_location(
  hnpi_interm,
  pdeath_filter,
  scenario_filter,
  hnpi_filter = c("incidC_shift"),
  fig_labs = c("Shift in case detection rate"),
  burn_in = 0
)
}
\arguments{
\item{burn_in}{number of iterations to discard before plotting}

\item{truth_dat}{df with date, geoid, incidI, incidDeath; hosps if adding
hospitalization data}

\item{model_dat}{df with model estimates}

\item{hosp}{whether hospitalization data is included in truth_dat with varname currhosp}

\item{filter_by}{variable name for filtering estimates either: scenario or pdeath}

\item{filter_val}{desired value of variable}

\item{geodat}{df with location names}
}
\value{
plot comparing observed and modeled estimates by geoid
}
\description{
Plot intermediate hnpi over time for each geoID
}
