% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_hosp_effec}
\alias{plot_hosp_effec}
\title{Comparison of pop-adjusted outcomes to effectiveness of current intervention}
\usage{
plot_hosp_effec(
  current_scenario,
  county_dat,
  start_date,
  end_date,
  geodat = geodata,
  r_dat = inference_r,
  pdeath_levels = c("high", "med", "low"),
  pdeath_labels = c("1\% IFR", "0.5\% IFR", "0.25\% IFR")
)
}
\arguments{
\item{current_scenario}{current scenario to plot}

\item{county_dat}{df with incident cases, hospitalizations, and deaths}

\item{start_date}{summarization period start}

\item{end_date}{summarization period end}

\item{geodat}{geodata file}

\item{r_dat}{df with effectiveness estimates, from load_r_sims_filtered}

\item{pdeath_levels}{death rate levels}

\item{pdeath_labels}{death rate labels}
}
\description{
Comparison of pop-adjusted outcomes to effectiveness of current intervention
}
