% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_hpar_by_location}
\alias{plot_hpar_by_location}
\title{Plot intermediate hpar over time for each geoID}
\usage{
plot_hpar_by_location(
  hpar_interm,
  pdeath_filter,
  scenario_filter,
  hpar_filter = c("incidC_probability", "incidD_probability"),
  fig_labs = c("Case detection probability", "Death probability"),
  burn_in = 0
)
}
\arguments{
\item{hpar_interm}{df with value, parameter, lik_type, scenario, pdeath, iter_num}

\item{pdeath_filter}{string that indicates which pdeath to import from outcome_dir}

\item{burn_in}{number of iterations to discard before plotting}

\item{partitions}{used by open_dataset}

\item{incl_geoids}{character vector of geoids that are included in the report}
}
\value{
plot comparing observed and modeled estimates by geoid
}
\description{
Plot intermediate hpar over time for each geoID
}
