% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_inference_r}
\alias{plot_inference_r}
\title{Plotting R or effectiveness estimates}
\usage{
plot_inference_r(
  r_dat,
  current_scenario,
  npi_labels = NA,
  npi_levels = NA,
  periodcolors = c("chartreuse3", "brown2", "turquoise4", "black"),
  effectiveness = FALSE,
  pi_lo = 0.25,
  pi_hi = 0.75,
  geodat = geodata
)
}
\arguments{
\item{r_dat}{df with R or reduction estimates per sim and npi to be included}

\item{current_scenario}{name of scenario inputs to use}

\item{npi_labels}{labels for plotted NPIs optional}

\item{npi_levels}{levels of NPIs optional}

\item{periodcolors}{}

\item{effectiveness}{whether to reduction estimates instead}

\item{pi_lo}{lower quantile for summarization}

\item{pi_hi}{higher quantile for summarization}

\item{geodat}{df with location names}
}
\value{
plot estimated R or effectiveness of intervention periods by geoid
}
\description{
Plotting R or effectiveness estimates
}
