% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_llik_by_location}
\alias{plot_llik_by_location}
\title{Plot intermediate likelihood over time for each geoID}
\usage{
plot_llik_by_location(
  llik_interm,
  filter_by = "pdeath",
  filter_val = "med",
  burn_in = 0
)
}
\arguments{
\item{llik_interm}{df with geoid, name, ll, lik_type, scenario, pdeath, iter_num}

\item{filter_by}{variable name for filtering estimates either: scenario or pdeath}

\item{filter_val}{desired value of variable}

\item{burn_in}{number of iterations to discard at the beginning}
}
\value{
plot of proposed and accepted log likelihood values by geoid
}
\description{
Plot intermediate likelihood over time for each geoID
}
