% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_model_vs_obs}
\alias{plot_model_vs_obs}
\title{Compare model outputs and data from CSSE}
\usage{
plot_model_vs_obs(
  state_hosp_totals,
  jhu_obs_dat,
  scenario_colors,
  pdeath_filter = pdeath_default,
  obs_data_col = "black",
  ci.L = 0,
  ci.U = 1,
  date_breaks = "1 month",
  sim_start_date,
  sim_end_date,
  week = FALSE,
  hosp = FALSE,
  tendency = "mean"
)
}
\arguments{
\item{state_hosp_totals}{df with model outputs summed over geoid (e.g., from load_hosp_geocombined_totals), with cols sim_num, pdeath, date, NincidCase, NincidDeath, NhospCurr}

\item{jhu_obs_dat}{df with observed data by geoid and cols incidI, incidDeath, and currhosp (if plotting hospitalizations)}

\item{scenario_colors}{character vector with scenario colors}

\item{pdeath_filter}{IFR level assumption}

\item{obs_data_col}{character string of observed data color}

\item{ci.L}{lower bound confidence interval}

\item{ci.U}{upper bound confidence interval}

\item{date_breaks}{breaks for dates in figure}

\item{sim_start_date}{simulation start date}

\item{sim_end_date}{simulation end date}

\item{week}{whether to aggregate incident cases/deaths to weeks (sum for incid cases/deaths, avg for daily hosp beds)}

\item{hosp}{whether to show hosp values}
}
\description{
Compare model outputs and data from CSSE
}
