% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_needs_relative_to_threshold_heatmap}
\alias{plot_needs_relative_to_threshold_heatmap}
\title{Plot heatmap of needs relative to a threshold (e.g. bed needs)}
\usage{
plot_needs_relative_to_threshold_heatmap(
  hosp_geounit_relative,
  scenario_labels,
  scale_colors = c("#066f6c", "#f8e6e7", "#ba0a0f"),
  legend_title,
  value_name,
  value_label,
  start_date,
  end_date,
  incl_geoids = NULL
)
}
\arguments{
\item{hosp_geounit_relative}{hosp_geounit_relative data}

\item{scenario_labels}{character vector of scenario labels from config}

\item{scale_colors}{character vector of colors for low, mid, and high plot values}

\item{legend_title}{label for legend}

\item{value_name}{name of secondary axis value}

\item{value_label}{secondary axis character label}

\item{start_date}{start date as character string "2020-01-01"}

\item{end_date}{end date as character string}

\item{shapefile}{object with geoid and name}
}
\value{
plot daily heatmap by geoid
}
\description{
Plot heatmap of needs relative to a threshold (e.g. bed needs)
}
