% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_outcome_rate}
\alias{plot_outcome_rate}
\title{Comparison of pop-adjusted outcomes by IFR}
\usage{
plot_outcome_rate(
  current_scenario,
  county_dat,
  start_date,
  end_date,
  geodat = geodata,
  pdeath_levels = c("high", "med", "low"),
  pdeath_labels = c("1\% IFR", "0.5\% IFR", "0.25\% IFR"),
  dodger = 0
)
}
\arguments{
\item{current_scenario}{current scenario}

\item{county_dat}{df with incident cases, hospitalizations, and deaths}

\item{start_date}{summarization period start}

\item{end_date}{summarization period end}

\item{geodat}{geodat file}

\item{pdeath_levels}{death rate levels}

\item{pdeath_labels}{death rate labels}

\item{dodger}{for plotting IFR estimates}
}
\description{
Comparison of pop-adjusted outcomes by IFR
}
