% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_scn_outcomes_ratio}
\alias{plot_scn_outcomes_ratio}
\title{Plot ratio of outcomes}
\usage{
plot_scn_outcomes_ratio(
  hosp_state_totals,
  start_date,
  end_date,
  pdeath_filter,
  scenario_colors,
  pi_lo,
  pi_hi
)
}
\arguments{
\item{hosp_state_totals}{df with hospitalization outcomes}

\item{start_date}{start of comparison period}

\item{end_date}{end of comparison period}

\item{pdeath_filter}{select pdeath: high, med, low}

\item{scenario_colors}{config$report$formatting$scenario_colors}

\item{pi_lo}{lower limit to interval}

\item{pi_hi}{upper lim to interval}
}
\value{
a table with the effectiveness per intervention period and a bar graph
}
\description{
Plot ratio of outcomes
}
