% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_spars}
\alias{plot_spars}
\title{Plot intermediate spar over time for each geoID}
\usage{
plot_spars(spar_interm, filter_by = "pdeath", filter_val = "med", burn_in = 0)
}
\arguments{
\item{spar_interm}{df with value, parameter, lik_type, scenario, pdeath, iter_num}

\item{filter_by}{variable name for filtering estimates either: scenario or pdeath}

\item{filter_val}{desired value of variable}

\item{burn_in}{number of iterations to discard before plotting}
}
\value{
plot comparing observed and modeled estimates by geoid
}
\description{
Plot intermediate spar over time for each geoID
}
