% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_truth_by_location}
\alias{plot_truth_by_location}
\title{Time series comparing reported and estimated cases and deaths in each
geoid by pdeath or scenario; possible to show hospitalization data}
\usage{
plot_truth_by_location(
  truth_dat,
  model_dat,
  hosp = FALSE,
  filter_by,
  filter_val,
  start_date,
  end_date,
  geodat = geodata,
  fig_labs = c("Incident Cases", "Incident Deaths"),
  pi_lo = 0.025,
  pi_hi = 0.975
)
}
\arguments{
\item{truth_dat}{df with date, geoid, incidI, incidDeath; hosps if adding
hospitalization data}

\item{model_dat}{df with model estimates}

\item{hosp}{whether hospitalization data is included in truth_dat with varname currhosp}

\item{filter_by}{variable name for filtering estimates either: scenario or pdeath}

\item{filter_val}{desired value of variable}

\item{geodat}{df with location names}

\item{fig_labs}{names for the labels of the figures that will show incidence cases, deaths, and hospitalizations}

\item{pi_lo}{lower limit to interval}

\item{pi_hi}{upper limit to interval}
}
\value{
plot comparing observed and modeled estimates by geoid
}
\description{
Time series comparing reported and estimated cases and deaths in each
geoid by pdeath or scenario; possible to show hospitalization data
}
