% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils-deprecated.R
\name{plot_ts_incid_ar_state}
\alias{plot_ts_incid_ar_state}
\title{Deprecated plot figure showing median and IQR time series for the state for one scenario}
\usage{
plot_ts_incid_ar_state(
  hosp_state_totals,
  geodata,
  pop_name,
  pdeath_level = "high",
  scenario_label,
  scenario_col,
  sim_start_date,
  sim_end_date,
  plot_intervention = FALSE,
  interv_start_date = NA,
  interv_end_date = NA
)
}
\arguments{
\item{hosp_state_totals}{totals for hospitalization related data for state for all pdeath}

\item{geodata}{geodata object from load_config}

\item{pdeath_level}{level of IFR (string: high/med/low) for filtering hospitalization data -- choose 1}

\item{scenario_label}{which scenario name from config}

\item{scenario_col}{which scenario color from config}

\item{sim_start_date}{simulation start date as character string "2020-01-01"}

\item{sim_end_date}{simulation end date as character string}

\item{plot_intervention}{logical indicating whether to plot grey box over a single intervention period -- will need to adapt if we want to show multiple intervention periods}

\item{interv_start_date}{intervention start date as character string}

\item{interv_end_date}{intervention end date as character string}
}
\value{
plot state time series median and IQR
}
\description{
Deprecated plot figure showing median and IQR time series for the state for one scenario
}
