% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{pretty_prop}
\alias{pretty_prop}
\title{Pretty print proportion for text}
\usage{
pretty_prop(x, y, markdown = TRUE)
}
\value{
a function that formats proportions
}
\description{
Pretty print proportion for text
}
