% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{print_pretty_date}
\alias{print_pretty_date}
\title{Pretty print date for text}
\usage{
print_pretty_date(date_string)
}
\value{
a function that formats full pretty dates
}
\description{
Pretty print date for text
}
