% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadFuncs-deprecated.R
\name{read_file_of_type}
\alias{read_file_of_type}
\title{Depracated function that returns a function to read files of a specific type (or automatically detected type based on extension)}
\usage{
read_file_of_type(extension, ...)
}
\arguments{
\item{extension}{The file extension to read files of}

\item{...}{Arguments to pass to the reading function}
}
\value{
A function which will read files with that extension.
 - We use readr::read_csv for csv files
 - We use arrow::read_parquet for parquet files
 - We use a function that detects the extension and calls this function if the auto extension is specified.
}
\description{
Depracated function that returns a function to read files of a specific type (or automatically detected type based on extension)
}
