% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{reference_chunk}
\alias{reference_chunk}
\title{Utility function for easily referencing chunks inluded as part of
the package as part of the report.}
\usage{
reference_chunk(chunkname)
}
\arguments{
\item{chunkname}{the file name of the chunk that is referenced}
}
\value{
the path to the chunk
}
\description{
Utility function for easily referencing chunks inluded as part of
the package as part of the report.
}
\author{
Justin Lessler
}
