% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{rt_estimates}
\alias{rt_estimates}
\title{Function to generate summary estimates of daily rt}
\usage{
rt_estimates(r_dat, county_dat, geodat, lo = 0.025, hi = 0.975)
}
\arguments{
\item{r_dat}{df with daily r estimates by geoid/sim}

\item{county_dat}{df with geoid, scenario, pdeath, sim_num, time, location, and cum_inf columns}

\item{lo}{}

\item{hi}{}
}
\description{
Function to generate summary estimates of daily rt
}
