% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{plot_rt_ts}
\alias{plot_rt_ts}
\title{Time series comparing Rt estimates by scenario over time}
\usage{
plot_rt_ts(
  rc,
  geodat,
  truth_dat,
  scenario_levels = scn_levels,
  scenario_labels = scn_labels,
  scenario_colors = scn_colors,
  truth_source = "JHU CSSE",
  pdeath_filter = "med"
)
}
\arguments{
\item{rc}{df with daily rt estimates for one geoid/location and columns "estimate", "lower", and "upper"}

\item{geodat}{df with geoid and population columns}

\item{scenario_levels}{levels applied to scenarios}

\item{scenario_labels}{label applied to scenarios}

\item{scenario_colors}{colors for each scenario}

\item{truth_source}{}

\item{pdeath_filter}{which pdeath to select}
}
\value{
a table with the effectiveness per intervention period and a bar graph
}
\description{
Time series comparing Rt estimates by scenario over time
}
