% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportBuildUtils.R
\name{rt_estimates}
\alias{rt_estimates}
\title{Function to generate summary estimates of daily rt}
\usage{
rt_estimates(
  r_dat,
  seir_dat,
  geodat,
  lo = 0.025,
  hi = 0.975,
  pdeath_filter = "med",
  overall = TRUE
)
}
\arguments{
\item{r_dat}{df with daily r estimates by geoid/sim}

\item{seir_dat}{df with geoid-specific outputs from SEIR}

\item{overall}{}
}
\description{
Function to generate summary estimates of daily rt
}
